requirements.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
celery==3.1.23
celerybeat-mongo==0.0.10
docutils==0.12
imagesize==0.7.1
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
mongoengine==0.10.6
Pygments==2.1.3
pymongo==3.2.2
pytz==2016.4
requests==2.9.1
six==1.10.0
slacker==0.9.10
snowballstemmer==1.2.1
Sphinx==1.4.1
websocket-client==0.37.0
